% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_watermark.R
\name{embed_lsb}
\alias{embed_lsb}
\title{Embed data in the least significant bits of an image}
\usage{
embed_lsb(img, message)
}
\arguments{
\item{img}{A 3D array (RGB image from png::readPNG)}

\item{message}{The string to embed}
}
\value{
The modified image array
}
\description{
Embed data in the least significant bits of an image
}
\keyword{internal}
